% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{EEZs-data}
\alias{EEZs-data}
\title{Western Pacific tropical Exclusive Economic Zone (EEZ) boundaries}
\format{
External GeoPackage file containing a SpatVector
}
\source{
\url{http://www.marineregions.org/downloads.php} (v10)
}
\description{
A spatial polygon data frame containing EEZs for Western Pacific tropical island nations.
This dataset is stored as an external file and can be loaded using the example code below.
}
\examples{
\dontrun{
library(terra)
eez_path <- system.file("extdata", "EEZs.gpkg", package = "VoCCdata")
if (file.exists(eez_path)) {
  EEZs <- terra::vect(eez_path)
  plot(EEZs)
}
}
}
