% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{HadiSST-data}
\alias{HadiSST-data}
\title{Global monthly mean sea surface temperatures (SST) for 1955-2010}
\format{
External GeoTIFF file containing a SpatRaster with 672 layers
}
\source{
Rayner et al. 2003. Global analyses of sea surface temperature, sea ice, and night marine air temperature since the late nineteenth century. J. Geophys. Res.Vol. 108: 4407. \url{https://www.metoffice.gov.uk/hadobs/hadisst/HadISST_paper.pdf}
}
\description{
A raster stack with 672 layers containing global mean monthly SSTs (HadISST 1.1) at 1-deg resolution for the period Jan 1955 to Dec 2010. Sourced from the Hadley Centre.
This dataset is stored as an external file and can be loaded using the example code below.
}
\examples{
\dontrun{
library(terra)
sst_path <- system.file("extdata", "HadiSST.tif", package = "VoCCdata")
if (file.exists(sst_path)) {
  HadiSST <- terra::rast(sst_path)
  plot(HadiSST[[1]])
}
}
}
\references{
\url{https://www.metoffice.gov.uk/hadobs/hadisst/}
}
