% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marshift}
\alias{marshift}
\title{Global metadataset of reported marine species range shifts}
\format{
A data frame with variables:
}
\source{
Poloczanska et al. 2013. Global imprint of climate change on marine life. Nat. Clim. Change, 3, 919-925. \url{https://www.nature.com/articles/nclimate1958}
}
\usage{
marshift
}
\description{
An extract from the meta data set used by Poloczanska et al.(2013) containing range shifts in marine taxa reported in the literature. Null responses removed and shifts reported as absolute shift distances.
}
\examples{
head(marshift)
}
\references{
Brown et al.(2016)
}
\keyword{datasets}
