% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traj25}
\alias{traj25}
\title{Climate velocity trajectories}
\format{
A data frame with variables:
}
\usage{
traj25
}
\description{
Climate velocity trajectories produced using the HSST data for a starting 0.25-degree coordinate grid and the period 1960-2009. The data frame contains coordinates and id for each trajectory.
}
\examples{
head(traj25)
}
\keyword{datasets}
